Moved `FeedImagesMapper` to dedicated file inside the Feed API group
